Client: code refactor
