+++ addOns/client/src/main/java/org/zaproxy/addon/client/internal/ReportedObject.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.client;
+package org.zaproxy.addon.client.internal;
 
 import java.util.Date;
 import net.sf.json.JSONObject;
