+++ addOns/client/src/main/java/org/zaproxy/addon/client/pscan/InformationInStorageScanRule.java
@@ -25,8 +25,8 @@ import net.sf.json.JSONObject;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.zaproxy.addon.client.ClientUtils;
-import org.zaproxy.addon.client.ReportedElement;
-import org.zaproxy.addon.client.ReportedObject;
+import org.zaproxy.addon.client.internal.ReportedElement;
+import org.zaproxy.addon.client.internal.ReportedObject;
 
 public class InformationInStorageScanRule extends ClientPassiveAbstractScanRule {
