+++ addOns/client/src/main/java/org/zaproxy/addon/client/pscan/JwtInStorageScanRule.java
@@ -28,8 +28,8 @@ import org.apache.commons.codec.binary.Hex;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.zaproxy.addon.client.ClientUtils;
-import org.zaproxy.addon.client.ReportedElement;
-import org.zaproxy.addon.client.ReportedObject;
+import org.zaproxy.addon.client.internal.ReportedElement;
+import org.zaproxy.addon.client.internal.ReportedObject;
 
 public class JwtInStorageScanRule extends ClientPassiveAbstractScanRule {
