+++ addOns/client/src/main/java/org/zaproxy/addon/client/ui/PopupMenuClientAttack.java
@@ -17,13 +17,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.client;
+package org.zaproxy.addon.client.ui;
 
 import java.awt.Component;
 import java.util.List;
 import javax.swing.JTree;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.extension.ExtensionPopupMenuItem;
+import org.zaproxy.addon.client.internal.ClientNode;
 import org.zaproxy.zap.extension.ExtensionPopupMenu;
 
 public class PopupMenuClientAttack extends ExtensionPopupMenu {
