+++ addOns/client/src/main/java/org/zaproxy/addon/client/ui/PopupMenuClientDelete.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.client;
+package org.zaproxy.addon.client.ui;
 
 import java.awt.event.ActionEvent;
 import javax.swing.JOptionPane;
