+++ addOns/client/src/main/java/org/zaproxy/addon/client/ui/PopupMenuItemClientHistory.java
@@ -17,37 +17,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.client;
+package org.zaproxy.addon.client.ui;
 
 import java.awt.Component;
 import java.awt.event.ActionEvent;
-import javax.swing.JTree;
 import org.parosproxy.paros.extension.ExtensionPopupMenuItem;
+import org.zaproxy.zap.view.ZapTable;
 
-public abstract class PopupMenuItemClient extends ExtensionPopupMenuItem {
+public abstract class PopupMenuItemClientHistory extends ExtensionPopupMenuItem {
 
     private static final long serialVersionUID = 1L;
-    private ClientMapPanel clientMapPanel;
+    private ClientHistoryPanel clientHistoryPanel;
 
-    public PopupMenuItemClient(String text, ClientMapPanel clientMapPanel) {
+    public PopupMenuItemClientHistory(String text, ClientHistoryPanel clientHistoryPanel) {
         super(text);
-        this.clientMapPanel = clientMapPanel;
+        this.clientHistoryPanel = clientHistoryPanel;
         this.addActionListener(l -> performAction(l));
     }
 
     @Override
     public boolean isEnableForComponent(Component invoker) {
-        if (invoker instanceof JTree) {
-            JTree tree = (JTree) invoker;
-            if (ClientMapPanel.CLIENT_TREE_NAME.equals(tree.getName())) {
-                return true;
-            }
-        }
-        return false;
+        boolean enabled =
+                invoker instanceof ZapTable
+                        && ClientHistoryPanel.CLIENT_HISTORY_NAME.equals(invoker.getName());
+        this.setEnabled(!clientHistoryPanel.getSelectedRows().isEmpty());
+        return enabled;
     }
 
-    public ClientMapPanel getClientMapPanel() {
-        return clientMapPanel;
+    public ClientHistoryPanel getClientHistoryPanel() {
+        return clientHistoryPanel;
     }
 
     abstract void performAction(ActionEvent e);
